backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
psycopg2-binary==2.9.9
python-docx==1.1.0
PyPDF2==3.0.1
langchain-text-splitters>=0.0.1
tiktoken==0.5.2
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
PyJWT>=2.8.0
